chapter_002/src/main/java/ru/job4j/pojo/Shop.java
package ru.job4j.pojo;

public class Shop {
    public static int indexOfNull(Product[] products) {
        for (int i = 0; i < products.length; i++) {
            if (products[i] == null) {
                return i;
            }
        }
        return -1;
    }

    public Product[] delete(Product[] products, int index) {
        products[index] = null;
        for (int i = index; i < products.length; i++) {
            if (i == products.length - 1) {
                products[products.length - 1] = null;
                break;
            }
            products[i] = products[i + 1];
        }
        return products;
    }
}
